Makes shodan.dts consistent with Renode platform
This was motivated by work on adding a timer driver, which will use the
device tree to define TIMER_PPTR. That does not quite work yet, as
documented in the new TODO in this change.

This change brings the whole .dts file up-to-date, which can stand on
its own to keep changes small.

Change-Id: I5b53d370a78ac50a724b94c507431c8c351e7387