Implements all seL4 timer methods in OpenTitan driver
Previously, we had hoped to rely on helper functions that convert
between ticks and microseconds from the arch timer.h. In practice that
was fragile (for example see the comment about integer overflow added in
this change).

Implementing all of the functions here will also allow excluding the
entire arch-level file rather than relying on #ifndef directives to
avoid multiple definitions.

This still will not actually be compiled as part of the build until the
timer is uncommented in shodan.dts, but that is still awaiting
a solution to the kernel device virtual address overflow problem.

Change-Id: I479d9aa97c26fdf7d3c39e9e4342dd4acd814d50