Stops clearing PLIC hart0 registers on seL4 boot
We should not clear PLIC registers for harts other than our own.
This was discovered when trying to debug a non-cleared interrupt on
the secure core. While this is not the cause, and while we may
ultimately use multiple PLICs, not touching registers for other targets
is still more correct.

Change-Id: Ib76b75b4262fe436e5938e845caeb30d25e8389b