kernel: add syscall to dump the contents of a CNode
Add a new syscall seL4_DebugDumpCNode that takes a seL4_CPtr and
outputs the capabilities in the CNode to the serial console.  If the
cptr is not for a CNode nothing is output to the console and an error
is returned.  This support requires CONFiG_PRINTING (but intentionally
not CONFIG_DEBUG_BUILD).

Change-Id: I3a0f5ff48a7f72964466c3c949edeb70b253b3b6