kernel: add nexus support
Update the platform definition config to finally break out to its own
CONFIG_PLAT_NEXUS. Unfortunately this was necessary to get the
additional memory/register layouts to be correctly specified in the
timer and uart drivers for riscv and will require retrofitting into
everything that keys off of CONFIG_PLAT_SHODAN in user space.

This includes just the timer and uart changes necessary for the
simulation. I'll bring in the vector interrupts in a follow-on CL.

Change-Id: I5f74fa2407422bdb94024a22c8570c57c9b0b0a6